Inline the common subdirectory. (#28)
Initially I created this subdir it because I thought it would be the
right structure for different flavors of the catalog. `common` was
supposed to be the base, and sibling directories would be created for
any variants necessary. It turned out the structure was somewhat
different, so there is no need for this.