Reorder attributes and add blank lines
I couldn't ever understand the logic of `opm` of sorting attributes
alphabetically. First you see a bunch of `entries` and only at the
end there's a `name` for them. That's not how yamls are normally
composed by humans. I changed that to have `name` fist, `schema`
second and `entries` last for `olm.channel`, for `olm.package` and
for the top-level element.